drop polyfill
Using a dynamic polyfill blocks a fully independent install. Also, this
polyfill does not buy much (in my opinion) when it comes to
compatibility:

- https://caniuse.com/intersectionobserver
- https://caniuse.com/es6

By dropping this, the primary "loss" is IE, of which all versions are
now officially EOL (and market share numbers are sub 1%).

This is well worth the trade-off IMO.