RF: starting point for new API with an entrypoint per command
This commit starts the process of introducing multiple
entrypoints by creating new and separate modules for
'create' and 'add'.

It also adds the following changes:
- introduces a dependency on datalad-next
- introduces the use of dl-next's constraint system
for parameter validation and coercion of new commands
- allows multiple formats for the metadata argument to
the 'add' command, including STDIN, stringified JSON,
and a file with JSON lines.
- introduces common fixtures via a new module and conftest
- updates the tests associated with 'create' and 'add'
- updates relevant test data
- adds entrypoints for 'create' and 'add' commands

The goal is for future commits adding more entrypoints
to build on top of the basis established here.