RF: changes location of catalog schemas
This commit is necessary to allow validation according to the schema
of a specific catalog, where previously validation was always done
according to the latest schema of the installed package version.

All catalog schemas now hava a dedicated location in
'catalog-name'/schema. If a metadata item is to be validated, this
location will be inspected in order to build up a schema store
that will be validated against. If this location does not exist,
the fallback is the default schema store of the package
installation.

Subsequent changes include:
- updating and moving packahe/catalog/tests/schema locations
to the constants module
- including the new schema location in package data
- adding methods to the webcatalog class to locate
the relevant schema store and validator
- updating create and add commands and their tests
accordingly