Refactoring of config functionality
This commit includes several changes that relate
to the refactoring of the config functionality. The
main changes relate to how/when a config file is provided.
Previously, config files were specified during
instantiation of the WebCatalog class, and a whole lot
of obscure code was necessary to determine how this config
applies to the catalog or dataset level. The refactored code
receives the config file via the Create() and Add() commands or via
the WebCatalog.create() or Node.add_attributes() python methods.

Important outcomes of the changed code include:
- WebCatalog can be instantiated with location alone (no 'action'
  necessary anymore, and no config_file)
- config is set on WebCatalog or Node instances after/during
  running their respective create() methods and not during
  instantiation
- WebCatalog class gained an 'add_record()' method to handle
  what was previously only done in Add() interface/command.
- MetaItem instantiation now handles config_file input, in order
  to pass it to Node.add_attributes() at the appropriate time.

Specific other changes include:
- splitting Node class into its own module
- updating utilities
- updating tests to conform to the refactors code