changes approach for setting the required tabs from URL parameters
This discard use of '' for finding available tabs
because of an issue with asynchronous loading of the
dataset component and its subcomponents, which include
tabs that might or might not be defined at the time
when 'created()' or 'mounted()' is called in the vue
app life cycle. This issue has in the past resulted
in erratic tab selection.

Instead, the new method determines the available tabs
from the root component data, and accesses this list
both on 'created()' and on 'beforeRouteUpdate()' to
check the provided URL parameter against.