Introduce back button for datasets that were navigated to inside the app
This introduces changes to html (to show and hide the back button),
to css (for the back button styling together with home button and dataset title),
and to JS. The latter handles logic for when to show and hide the back button.
Upon 'create()' the superdataset file is fetched in order to store the id and version
on the component, and the back button is hidden. Upon internal navigation, the JS
checks if the route that is navigated to is the same as the home page by comparing
the new id and version to the stored super id and version. If they are different,
show the back button; if they are the same, hide the back button.