update display and behaviour of items in the content tab filetree
For items of type file, the cloud icon is removed and the filename is turned into a link
(if the file has an associated URL), and the filesize is updated to display correctly
(empty string and not NaN, if the value is empty). Also, for all items, the cursor will
only turn into a pointer if the item is a folder or a file with a link.