adds a datalad-catalog contained extractor for datalad datasets
This ports most functionality from datalad_metalad.extractors.core into a script
that also then adds translation into the catalog schema. The script receives
a path to a datalad dataset as parameter and outputs a metadata record that can
immediatelt be added toa catalog. In this way, the dependence on metalad is
removed and the explicit Translator functionality of the catalog (which also
depends on jq bindings) does not have to be used. The reason for doing this is
to have a self-contained script that could in future just be ripped and replaced
with whatever new functionality supercedes this.