Add new component to render an 'additional_tab_display' with context information
This update stems from the work done on the SFB1451 catalog regarding
the rendering of semantic information, see:
https://github.com/psychoinformatics-de/sfb1451-projects-catalog/pull/62/files.
A new async component is introduced to encapsulate any rendering of 'additional_tab_display's
that have an '@context' key in their 'content' field. The component was created to remove
specific implementation details related to semantics from the overall dataset page rendering
(i.e. from the dataset-template component). This allows for future changes to the semantic tab
rendering without affecting how the component is invoked.