Adds functionality to summarize info about dataset ids and versions in a catalog
This includes changes to the following core modules:
- webcatalog: new methods are added for extracting all IDs
in a catalog, and all VERSIONs for respective IDs, including
properties like name, alias, and more. This allows users to
iterate through a list comprising all catalog datasets (which
is necessary for the upcoming 'tree' command) and allows the
'report' functionality which summarizes some stats of a catalog.
- node: adds a method for sorting metadata_sources by time, as
a proxy for determining a 'last_updated' value for a dataset
version, which forms part of the output of the new webcatalog
methods.
- get: adds the property 'report', allowing a summary report
of the catalog to be printed out.
- utils: changes supporting above methods.
- tests: minor changes to allow test success, new tests are TODO.

These commits all contribute to the upcoming changes related
to navigating to a dataset-version via concept ID or via alias.