updates to javascript assets to allow user-friendly navigation
These are the web-side changes that accompany the previous commits
related to dataset aliases and dataset id concepts. This now allows
users to navigate to e.g. http://mycatalog.de/dataset/\<dataset-id\>
or to e.g. http://mycatalog.de/dataset/\<dataset-alias\>. The javascript
will first fetch the concept id file or the alias file, which contains a
redirect pointer to the correct dataset id AND version. Internal navigation
between datasets are still done based on the dataset-id AND dataset-version
URL parameters. The latter now becomes an optional parameter. One aspect
that this commit does not account for yet is the second optional URL parameter
'tab' and its use with the sinle required URL parameter.