creates a runprov exctractor, ported from datalad-metalad
this commit adds a script that runs a slightly refactored version of the 'runprov'
extractor in datalad-metalad. Additionally, it translates the output of that code
into a metadata record that is compliant witht datalad-catalog's dataset schema,
such that the script's output can be directly 'catalog-added' as an entry to
an existing catalog. The main reason for porting this functionality is to have a
self-contained script inside the package that makes dependence on metalad unnecessary.