wip: work on instating tabs and keywords as route query parameters
This falls into the same context as the work on aliasing and concept id
navigation. The point is that any number of combinations between using
an alias or concept id or dataset_id+version, and query tabs and keywords
as query parameters should be possible. An important change is the use of
history.replaceState to update the query string in the URL without updating
the route or reloading it (i.e. without using pushState). This seems to work
fine, but the process of keeping the query string and the selected keywords
and tabs in sync is still not ironed out completely, and hence some navigations
are still a bit buggy (e.g. after navigation to a subdataset, a keyword query
param might still remain, unexpectedly). Note that vue-router's router.push
is not aware of the history.replaceState call. This is not a problem per se,

but needs to be kept track of.