More functionality to consolidate router.push and history.replaceState behaviour
This commit fixes bugs in the recently added functionality to update the URL query
string based on user selections. The problems included that a test for '>= 0'
incorrectly validated null or NaN, resulting in the incorrect part of an if statement
being executed and creating the '?tab=undefined' issue in the query string; and also
that the code for adding/removing search tags based on the 'keyword' parameter in the
query string had to move to the 'dataset_ready' watcher rather than the 'subdatasets_ready'
watcher to allow more time for asynchronous code to complete. To support debugging
during this and future steps, several 'console.debug' statements were added all
over the code base. Lastly, references in the html rendering to 'Subdatasets' were
replaced with 'Datasets' to conform to recent similar changes in other catalog instances.