Increase the richness of structured data for google dataset search
The goal here is to not only provide the minimum required fields to make
a catalog dataset discoverable by google dataset search (this is already
possible) but to now also include any optional fields to enrich the
structured data for the dataset when viewing it in a dataset search
result.

The changes necessary to allow this include:
- basic functions to map all optional fields from existing properties
in a metadata entry
- an update to the catalog-level config file to include fields for
'catalog_name' and 'catalog_url' that are used for mapping to the
optional 'includedInDataCatalog' rich data field

Any existing or new catalog that wants to make use of these features
should ensure that their catalog-level config file is updated.