Fix openWithBinder handling of custom notebook URLs
After the ability to modify content url, repo name, or notebook name
was introduced in 5ac68d7, these can no longer be const, and need to
be var.

Otherwise, clicking the "Explore with Binder" button would log a
"TypeError: assignment to constant variable" in the console and do
nothing else if custom notebook information was present.