Add caching headers to fetch calls that get json metadata
These headers specify 'cache: no-cache' so that the fetch call will
not rely on cached metadata which could be outdated. This is
especially needed when navigating to a catalog's homepage, which
which fetches the catalog's 'super.json' file to look for the home
page dataset to display, which could change often.

In addition, the outdated XMLHttpRequest call is replaced with a fetch