fix circular import error with loaded datalad-next
This commit fixes a circular import error that
occurs when datalad-next is loaded and tests are
patched and the doctest-module of pytest is
activated. The error is described in
<datalad/datalad-next#716>

The reason for the import error is that importing
datalad_catalog.add from the module-level context
in datalad_catalog.tests.test_add, will import
datalad_next, which will in turn import certain
tests, which will trigger the extension-api
generation process, which tries to process
datalad_catalog.add, which is not yet
completely loaded.

This commit disables the doctest-modulei. This
prevents the circular import failure that
originates in the import of datalad_next.add