Introduce show/hide flags for dataset buttons
The visibility of dataset buttons have until now depended mostly on
inherent logic, such as whether a specific property related to the button
is included in the dataset-level metadata or not. This commit introduces
controlling flags that users can specify via catalog- or dataset-level
config files. These flags can explicitly be set to false if the user
prefers to hide specific buttons, while the default for these flags is
true (i.e. show the buttons, unless some data-related logic specifies not to).

The updated template config file at 'datalad_catalog/config/config.json' contains
all new controlling flags.

In addition to the code changes to make the above possible, the way in which
the 'dataset_options' are set for a selected dataset has been updated.
dataset_options should be an object that is a merge between the catalog-level
config and the dataset-level config, with dataset-level taking priority.